Add variable age order optimization to unify
Only helps a little vs the default order given the way we usually
write unifications, but it makes unification performance less
dependent on argument order. (It is significantly faster than
unification with the opposite-of-default argument order).

Jason found the idea in "ON THE COMPLEXITY OF UNIFICATION SEQUENCES"
by Heikki Mannila and Esko Ukkonen. Lemma 6, pg 130.