Provide succeed and fail.
`succeed` and `fail` are used in The Reasoned Schemer, it seems reasonable to provide them. The book also uses the reader syntax `#s` and `#u`, which might be interesting to provide but that would require a reader extension and probably a `#lang`. That doesn't seem worth the effort at the moment.